requirements.txt
# ─── Database ───
psycopg2-binary>=2.9.9

# ─── Image handling (OCR tool) ───
Pillow>=10.0.0

# ─── Telegram webhook harness (dev/test) ───
flask>=3.0.0
requests>=2.31.0
